Limpiar y unificar bots
