+++ .gitignore
@@ -1,2 +1,1 @@
-tiktok/tiktok
-instagram/instagram
+dlbot
